+++ build.rs
@@ -1,4 +1,4 @@
 fn main() {
-    peg::cargo_build("src/config_grammar.rustpeg");
-    peg::cargo_build("src/preprocess_grammar.rustpeg");
+    peg::cargo_build("src/grammars/config.rustpeg");
+    peg::cargo_build("src/grammars/preprocess.rustpeg");
 }
