+++ src/io.rs
@@ -1,6 +1,7 @@
-use std::fs::{File};
 use std::io;
-use std::io::{Read, Seek, Write, Stdout, Cursor};
+use std::io::{Cursor, Error, Read, Seek, Stdout, Write};
+use std::fs::File;
+use std::path::PathBuf;
 
 pub enum Input {
     File(File),
@@ -121,4 +122,42 @@ pub fn compressed_int_len(x: u32) -> usize {
     }
 
     len + 1
+}
+
+pub fn matches_glob(s: &String, pattern: &String) -> bool {
+    if let Some(index) = pattern.find('*') {
+        if s[..index] != pattern[..index] { return false; }
+
+        for i in (index+1)..(s.len()-1) {
+            if matches_glob(&s[i..].to_string(), &pattern[(index+1)..].to_string()) { return true; }
+        }
+
+        false
+    } else {
+        s == pattern
+    }
+}
+
+pub fn file_allowed(name: &String, exclude_patterns: &[String]) -> bool {
+    for pattern in exclude_patterns {
+        if matches_glob(&name, &pattern) { return false; }
+    }
+    true
+}
+
+pub fn list_files(directory: &PathBuf) -> Result<Vec<PathBuf>, Error> {
+    let mut files: Vec<PathBuf> = Vec::new();
+
+    for entry in std::fs::read_dir(directory)? {
+        let path = entry?.path();
+        if path.is_dir() {
+            for f in list_files(&path)? {
+                files.push(f);
+            }
+        } else {
+            files.push(path);
+        }
+    }
+
+    Ok(files)
 }
