+++ src/lib.rs
@@ -5,7 +5,8 @@ pub mod config;
 pub mod error;
 pub mod io;
 pub mod p3d;
-pub mod pbo;
 pub mod preprocess;
 pub mod run;
 pub mod sign;
+
+pub mod pbo;
